Encode inline script data through the rendering helpers

+++ user_group_admin.php
@@ -1678,7 +1678,7 @@ function user_group_settings_edit(string $header_label) : void {
 	?>
 	<script type='text/javascript'>
 
-	var themeFonts=<?php print read_config_option('font_method'); ?>;
+	var themeFonts=<?php print (int) read_config_option('font_method'); ?>;
 
 	function graphSettings() {
 		var showField = $('#custom_fonts').is(':checked');
